Update - moved project-images to public
