+++ my-react-portfolio/public/project-images/project-library-solo.jpeg
