--- my-react-portfolio/src/assets/project-images/project-images.json
@@ -1,30 +1,0 @@
-[
-  {
-    "repoName": "Project-Happy-Thoughts-Vite",
-    "imageUrl": "/src/assets/project-images/project-happy-thoughts.jpeg"
-  },
-  {
-    "repoName": "project-chatbot",
-    "imageUrl": "/src/assets/project-images/project-chatbot.jpeg"
-  },
-  {
-    "repoName": "project-library-solo",
-    "imageUrl": "/src/assets/project-images/project-library-solo.jpeg"
-  },
-  {
-    "repoName": "project-music-releases-vite",
-    "imageUrl": "7src/assets/project-images/project-music-releases-vite.jpeg"
-  },
-  {
-    "repoName": "project-survey-vite-solo",
-    "imageUrl": "/src/assets/project-images/project-survey-vite-solo.jpeg"
-  },
-  {
-    "repoName": "project-weather-app",
-    "imageUrl": "/src/assets/project-images/project-weather-app.jpeg"
-  },
-  {
-    "repoName": "project-news-site",
-    "imageUrl": "/src/assets/project-images/project-news-site.jpeg"
-  }
-]
