+++ my-react-portfolio/src/components/projects/ProjectImage.jsx
@@ -1,5 +1,5 @@
 import "./projectImage.css"
-import projectImages from "../../assets/project-images/project-images.json"
+import projectImages from "../../project-images.json"
 
 export const ProjectImage = ({repo}) => {
   const repoImage = projectImages.find(img => img.repoName === repo);
